src/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
distlib==0.3.8
Django==5.0.4
django-extensions==3.2.3
django-filter==24.2
djangorestframework==3.15.1
filelock==3.13.4
pillow==10.3.0
pipenv==2023.12.1
platformdirs==4.2.1
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.11.0
tzdata==2024.1
virtualenv==20.26.0
